Move common acknowledgement handling to AbstractHandle; sort branches by specificity; remove unreachable branch; add a test for feature acknowledgement requests.
